+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/extensions/FlowExt.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.util
+package com.woocommerce.android.extensions
 
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
