+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/login/ObserveLoginRequest.kt
@@ -1,25 +1,22 @@
 package com.woocommerce.android.ui.login
 
+import com.woocommerce.android.extensions.combineWithTimeout
 import com.woocommerce.android.ui.login.ObserveLoginRequest.LoginRequestState.Logged
 import com.woocommerce.android.ui.login.ObserveLoginRequest.LoginRequestState.Timeout
 import com.woocommerce.android.ui.login.ObserveLoginRequest.LoginRequestState.Waiting
-import com.woocommerce.android.util.timeoutFlow
-import kotlinx.coroutines.flow.combine
 import javax.inject.Inject
 
 class ObserveLoginRequest @Inject constructor(
     private val loginRepository: LoginRepository
 ) {
-    operator fun invoke() = combine(
-        loginRepository.isSiteAvailable,
-        timeoutFlow
-    ) { isSiteAvailable, isTimeout ->
-        when {
-            isSiteAvailable -> Logged
-            isTimeout.not() -> Waiting
-            else -> Timeout
+    operator fun invoke() = loginRepository.isSiteAvailable
+        .combineWithTimeout { isSiteAvailable, isTimeout ->
+            when {
+                isSiteAvailable -> Logged
+                isTimeout.not() -> Waiting
+                else -> Timeout
+            }
         }
-    }
 
     enum class LoginRequestState {
         Logged,
