+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/ui/stats/datasource/FetchStatsFromPhone.kt
@@ -1,7 +1,7 @@
 package com.woocommerce.android.ui.stats.datasource
 
+import com.woocommerce.android.extensions.combineWithTimeout
 import com.woocommerce.android.phone.PhoneConnectionRepository
-import com.woocommerce.android.util.combineWithTimeout
 import com.woocommerce.commons.wear.MessagePath.REQUEST_STATS
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.filterNotNull
@@ -13,12 +13,13 @@ class FetchStatsFromPhone @Inject constructor(
 ) {
     suspend operator fun invoke(): Flow<StoreStatsRequest?> {
         phoneRepository.sendMessage(REQUEST_STATS)
-        return statsRepository.observeStatsDataChanges().combineWithTimeout { statsData, isTimeout ->
-            when {
-                statsData?.isFinished == true -> statsData
-                isTimeout -> StoreStatsRequest.Error
-                else -> null
-            }
-        }.filterNotNull()
+        return statsRepository.observeStatsDataChanges()
+            .combineWithTimeout { statsData, isTimeout ->
+                when {
+                    statsData?.isFinished == true -> statsData
+                    isTimeout -> StoreStatsRequest.Error
+                    else -> null
+                }
+            }.filterNotNull()
     }
 }
